requirements.txt
pyPDF
tkinter